pyproject.toml
[project]
name = "assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "decouple>=0.0.7",
    "openai-agents>=0.2.2",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.11",
]
